some changes
1. replace integer instead of bigdecimal
2. add github actions
3. add linter
4. config unit test files correctly

+++ src/main/java/conekta/io/model/submodel/Charge.java
@@ -3,19 +3,18 @@ package conekta.io.model.submodel;
 import conekta.io.model.ConektaObject;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 public class Charge extends ConektaObject {
     private String status;
-    private BigDecimal amount;
-    private BigDecimal fee;
+    private Integer amount;
+    private Integer fee;
     private String orderId;
     private String customerId;
     private Boolean livemode;
-    private BigDecimal monthlyInstallments;
-    private BigDecimal createdAt;
-    private BigDecimal paidAt;
+    private Integer monthlyInstallments;
+    private Integer createdAt;
+    private Integer paidAt;
     private String currency;
     private String description;
     private PaymentMethod paymentMethod;
