+++ src/main/java/conekta/io/model/submodel/Checkout.java
@@ -3,7 +3,6 @@ package conekta.io.model.submodel;
 import conekta.io.model.ConektaObject;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,12 +11,12 @@ public class Checkout extends ConektaObject {
     private List<Integer> monthlyInstallmentsOptions;
     private List<String> allowedPaymentMethods;
     private List<Object> excludeCardNetworks;
-    private BigDecimal emailsSent;
+    private Integer emailsSent;
     private String successUrl;
     private String failureUrl;
-    private BigDecimal expiredAt;
-    private BigDecimal expiresAt;
-    private BigDecimal startsAt;
+    private Integer expiredAt;
+    private Integer expiresAt;
+    private Integer startsAt;
     private Boolean force3dsFlow;
     private Boolean livemode;
     private Boolean monthlyInstallmentsEnabled;
@@ -25,10 +24,10 @@ public class Checkout extends ConektaObject {
     private Boolean canNotExpire;
     private String name;
     private Boolean needsShippingContact;
-    private BigDecimal paidPaymentsCount;
+    private Integer paidPaymentsCount;
     private Boolean recurrent;
     private String slug;
-    private BigDecimal smsSent;
+    private Integer smsSent;
     private String status;
     private String type;
     private String url;
