+++ src/main/java/conekta/io/model/submodel/LineItem.java
@@ -2,13 +2,12 @@ package conekta.io.model.submodel;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class LineItem extends ConektaData {
     private String name;
-    private BigDecimal unitPrice;
+    private Integer unitPrice;
     private Integer quantity;
     private List<String> tags;
     private Object metadata;
