+++ src/main/java/conekta/io/model/submodel/PaymentMethod.java
@@ -4,13 +4,12 @@ import conekta.io.model.ConektaObject;
 import lombok.Data;
 import lombok.ToString;
 
-import java.math.BigDecimal;
 
 @Data
 public class PaymentMethod extends ConektaObject {
 
     private String type;
-    private BigDecimal createdAt;
+    private Integer createdAt;
     private String last4;
     private String bin;
     private String expMonth;
@@ -24,12 +23,12 @@ public class PaymentMethod extends ConektaObject {
     private String storeName;
     private String serviceName;
     private String serviceNumber;
-    private BigDecimal expiresAt;
+    private Integer expiresAt;
     private String description;
     private String authCode;
     private String store;
     private DefaultAddress address;
     private String tokenID;
     private String paymentSourceID;
-    private BigDecimal monthlyInstallments;
+    private Integer monthlyInstallments;
 }
